+++ R/addDepot.R
@@ -1,10 +1,15 @@
-#' To convert from Infusion/intravenous administration to first order oral absorption
-#' @param model The model as a function
+#' To convert from infusion/intravenous administration to first-order oral
+#' absorption
+#'
+#' @param model The model as a function (or something convertible to an rxUi
+#'   object)
 #' @param central central compartment name
-#' @param depot depot name
-#' @param absRate absorption rate
-#' @param lag A boolean representing if you are going to add a lag to this compartment
-#' @param tlag a character vector representing the lag time
+#' @param depot depot compartment name
+#' @param absRate absorption rate parameter name
+#' @param lag lag parameter name
+#' @param lagIni Initial value for the lag time (`NA` to omit)
+#' @param fdepotIni Initial value for the depot bioavailability (`NA` to omit)
+#' @param absRateIni Initial value for the first order rate
 #' @return a model with the depot added
 #' @export
 #' @examples
@@ -13,12 +18,25 @@
 #' readModelDb("PK_1cmt_des") |>
 #'   removeDepot() |>
 #'   addDepot()
-addDepot <- function(model, central = "central", depot = "depot", absRate = "ka", lag = FALSE, tlag = "lagD") {
-  checkmate::assertCharacter(depot, pattern = "^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$", len = 1, any.missing = FALSE, min.chars = 1)
-  checkmate::assertCharacter(absRate, pattern = "^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$", len = 1, any.missing = FALSE, min.chars = 1)
-  checkmate::assertLogical(lag, len = 1, any.missing = FALSE)
-  checkmate::assertCharacter(tlag, pattern = "^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$", len = 1, any.missing = FALSE, min.chars = 1)
+addDepot <- function(model,
+                     central = "central", depot = "depot",
+                     absRate = "ka", lag = paste0("lag", depot),
+                     lagIni=NA, fdepotIni=NA,
+                     absRateIni=1.0) {
+  model <- rxode2::assertRxUi(model)
+  assertCompartmentName(depot)
+  assertCompartmentExists(model, central)
+  assertVariableName(absRate)
+  assertVariableName(lag)
+  if (!is.na(lagIni)) {
+    assertParameterValue(lagIni)
+  }
+  if (!is.na(fdepotIni)) {
+    assertParameterValue(fdepotIni)
+  }
+  assertParameterValue(absRateIni)
   temp <- rxode2::assertRxUi(model)
+
   mv <- rxode2::rxModelVars(temp)
   if (absRate %in% mv$params) {
     stop("'", absRate, "' cannot be in the model")
@@ -47,45 +65,131 @@ addDepot <- function(model, central = "central", depot = "depot", absRate = "ka"
 
   # Additional equations to be added to model block
   absrateModel <- paste0(absRate, " <- exp(l", absRate, ")")
-  fdepotModel <- paste0("f", depot, " <- exp(lf", depot, ")")
-  lagModel <- paste0(tlag, " <- exp(la", tlag, ")")
-  fdepotODE <- paste0("f(", depot, ") <- f", depot)
-  depotODE <- paste0("d/dt(", depot, ") <- -", absRate, "*", depot)
-  lagODE <- paste0("alag(", depot, ") <- ", tlag)
-
-  # Modify model block
-  if (lag == FALSE) {
-    rxode2::model(temp) <- c(absrateModel, fdepotModel, model[1:(centralLine - 1)], depotODE, fdepotODE, line, model[(centralLine + 1):length(model)])
+  
+  # Lag equations
+  if (is.na(lagIni)) {
+    lagModel <- lagODE <- NULL
   } else {
-    rxode2::model(temp) <- c(absrateModel, fdepotModel, lagModel, model[1:(centralLine - 1)], depotODE, fdepotODE, lagODE, line, model[(centralLine + 1):length(model)])
+    lagModel <- paste0(lag, " <- exp(la", lag, ")")
+    lagODE <- paste0("alag(", depot, ") <- ", lag)
+  }
+  
+  # Bioavailability equations
+  if (is.na(fdepotIni)) {
+    fdepotModel <- fdepotODE <- NULL
+  } else {
+    fdepotModel <- paste0("f", depot, " <- exp(lf", depot, ")")
+    fdepotODE <- paste0("f(", depot, ") <- f", depot)
   }
 
+  depotODE <- paste0("d/dt(", depot, ") <- -", absRate, "*", depot)
+
+  # Modify model block
   if (any(grepl("^transit", mv$state))) {
-    rxode2::model(temp) <- c(absrateModel, fdepotModel, lagModel, model[1:(transitLine - 1)], depotODE, fdepotODE, lagODE, transitODE, model[(transitLine + 1):(centralLine - 1)], lineNew, model[(centralLine + 1):length(model)])
+    rxode2::model(temp) <-
+      c(
+        absrateModel,
+        fdepotModel,
+        lagModel,
+        model[1:(transitLine - 1)],
+        depotODE,
+        fdepotODE,
+        lagODE,
+        transitODE,
+        model[(transitLine + 1):(centralLine - 1)],
+        lineNew,
+        model[(centralLine + 1):length(model)]
+      )
+  } else {
+    rxode2::model(temp) <-
+      c(
+        absrateModel,
+        fdepotModel,
+        lagModel,
+        model[1:(centralLine - 1)],
+        depotODE,
+        fdepotODE,
+        lagODE,
+        line,
+        model[(centralLine + 1):length(model)]
+      )
   }
 
   # Modify ini block
-  rateIni <- str2lang(paste0("l", absRate, " <-0.02"))
-  lfdepotIni <- str2lang(paste0("lf", depot, " <-0.04"))
-  lalagIni <- str2lang(paste0("la", tlag, " <-0.09"))
-  if (lag == FALSE) {
-    temp <- rxode2::ini(temp, rateIni, append = 0) |>
-      rxode2::ini(lfdepotIni)
+  rateIni <- str2lang(paste0("l", absRate, " <-", absRateIni))
+  lfdepotIni <- str2lang(paste0("lf", depot, " <-", fdepotIni))
+  if (is.na(lagIni)) {
+    temp <- rxode2::ini(temp, rateIni, append = 0)
+    if (!is.na(fdepotIni)) {
+      temp <- rxode2::ini(temp, lfdepotIni)
+    }
     temp2 <- temp$iniDf
     suppressMessages(temp2$label[temp2$name == paste0("l", absRate)] <- paste0("First order absorption rate (", absRate, ")"))
     suppressMessages(temp2$label[temp2$name == paste0("lf", depot)] <- "Bioavailability (F)")
   } else {
+    lalagIni <- str2lang(paste0("la", lag, " <-", lagIni))
     temp <- temp |>
       rxode2::ini(rateIni, append = 0) |>
-      rxode2::ini(lfdepotIni) |>
       rxode2::ini(lalagIni)
+    if (!is.na(fdepotIni)) {
+      temp <- rxode2::ini(temp, lfdepotIni)
+    }
     temp2 <- temp$iniDf
     suppressMessages(temp2$label[temp2$name == paste0("l", absRate)] <- paste0("First order absorption rate (", absRate, ")"))
-    suppressMessages(temp2$label[temp2$name == paste0("lf", depot)] <- "Bioavailability (F)")
-    suppressMessages(temp2$label[temp2$name == paste0("la", tlag)] <- paste0("Lag time (", tlag, ")"))
+    if (!is.na(fdepotIni)) {
+      suppressMessages(temp2$label[temp2$name == paste0("lf", depot)] <- "Bioavailability (F)")
+    }
+    suppressMessages(temp2$label[temp2$name == paste0("la", lag)] <- paste0("Lag time (", lag, ")"))
   }
   rxode2::ini(temp) <- temp2
 
   # return
+  temp
+}
+
+#' To convert from first order oral absorption to IV/Intravenous
+#' @param model The model as a function
+#' @param central This is a character vector that represents the central compartment in the model
+#' @param depot  This is a character vector that represents the depot in the model
+#' @return Returns a model with the depot from a first order absorption model removed
+#' @export
+#' @examples
+#' readModelDb("PK_1cmt_des") |>
+#'   removeDepot()
+removeDepot <- function(model, central = "central", depot = "depot") {
+  assertCompartmentName(central)
+  assertCompartmentName(depot)
+  
+  temp <- rxode2::assertRxUi(model)
+  mv <- rxode2::rxModelVars(temp)
+  if (!(central %in% mv$state)) {
+    stop("'", central, "' needs to be in the model")
+  }
+  if (!(depot %in% mv$state)) {
+    stop("'", depot, "' needs to be in the model")
+  }
+  if (any(grepl("^transit", mv$state))) {
+    transit <- eval(str2lang(paste0("rxode2::modelExtract(temp,d/dt(transit1),lines=TRUE)")))
+    transitLine <- attr(transit, "lines")
+    transitNew <- str2lang(sub("\\s*ka\\s*\\*\\s*depot", "", transit))
+  }
+  
+  
+  model <- rxode2::modelExtract(temp, endpoint = NA)
+  center <- eval(str2lang(paste0("rxode2::modelExtract(temp,d/dt(", central, "),lines=TRUE)")))
+  centralLine <- attr(center, "lines")
+  rhs <- str2lang(sub("\\s*ka\\s*\\*\\s*depot", "", center))
+  if (any(grepl("^transit", mv$state))) {
+    rxode2::model(temp) <- c(model[1:(transitLine - 1)], transitNew, model[(transitLine + 1):(centralLine - 1)], rhs, model[(centralLine + 1):length(model)])
+  } else {
+    rxode2::model(temp) <- c(model[1:(centralLine - 1)], rhs, model[(centralLine + 1):length(model)])
+  }
+  
+  ka <- fdepot <- depot <- d <- dt <- f <- NULL
+  if ("fdepot" %in% mv$lhs) {
+    temp <- rxode2::model(temp, -ka, -fdepot, -f(depot), -d / dt(depot))
+  } else {
+    temp <- rxode2::model(temp, -ka, -d / dt(depot))
+  }
   temp
 }